+++ src/main/java/dk/dataforsyningen/arkivmeta/apimodel/FaeroesketopokortDto.java
@@ -1,15 +1,16 @@
 package dk.dataforsyningen.arkivmeta.apimodel;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public class FaeroesketopokortDto extends KortDto {
   @Schema(description = "Det år den oprindelige opmåling til kortet blev afsluttet. Kortet kan senere være nymålt eller rettet.")
-  private Integer aarformaalt;
+  private BigDecimal aarformaalt;
 
   @Schema(description = "Det år kortet blev udgivet eller trykt. Kortene er udgivet efter deres opmåling og eventuelt genudgivet som følge af rettelser eller at kortet blev udsolgt.")
-  private Integer aarforudgivelse;
+  private BigDecimal aarforudgivelse;
 
   @Schema(description = "Versionsnummer for kortet. Et kort kan være udgivet i flere versioner.")
   private String version;
@@ -23,8 +24,8 @@ public class FaeroesketopokortDto extends KortDto {
                               String bemaerkning,
                               List<String> daekningsomraade,
                               List<String> filer,
-                              Integer gaeldendefra,
-                              Integer gaeldendetil,
+                              BigDecimal gaeldendefra,
+                              BigDecimal gaeldendetil,
                               String geometri,
                               String kortbladnummer,
                               String kortvaerk,
@@ -35,8 +36,8 @@ public class FaeroesketopokortDto extends KortDto {
                               LocalDateTime registreringtil,
                               String titel,
                               String uniktkortnavn,
-                              Integer aarformaalt,
-                              Integer aarforudgivelse,
+                              BigDecimal aarformaalt,
+                              BigDecimal aarforudgivelse,
                               String version) {
     super(id,
         alternativtitel,
@@ -61,19 +62,19 @@ public class FaeroesketopokortDto extends KortDto {
     this.version = version;
   }
 
-  public Integer getAarformaalt() {
+  public BigDecimal getAarformaalt() {
     return aarformaalt;
   }
 
-  public void setAarformaalt(Integer aarformaalt) {
+  public void setAarformaalt(BigDecimal aarformaalt) {
     this.aarformaalt = aarformaalt;
   }
 
-  public Integer getAarforudgivelse() {
+  public BigDecimal getAarforudgivelse() {
     return aarforudgivelse;
   }
 
-  public void setAarforudgivelse(Integer aarforudgivelse) {
+  public void setAarforudgivelse(BigDecimal aarforudgivelse) {
     this.aarforudgivelse = aarforudgivelse;
   }
